+++ src/pint/observatory/topo_obs.py
@@ -19,10 +19,6 @@ from pint.utils import has_astropy_unit
 
 # These are global because they are, well, literally global
 _gps_clock = None
-# have the IERS-B data been downloaded?
-_IERSB_download = False
-# has the lead-second file been downloaded?
-_leapsecond_download = False
 
 
 class TopoObs(Observatory):
@@ -240,20 +236,6 @@ class TopoObs(Observatory):
     def earth_location_itrf(self, time=None):
         return self._loc_itrf
 
-    def _load_iersb(self):
-        global _IERSB_download
-        if not _IERSB_download:
-            log.debug("Running get_iers_up_to_date() to update IERS B table")
-            get_iers_up_to_date()
-            _IERSB_download = True
-
-    def _load_leapsecond(self):
-        global _leapsecond_download
-        if not _leapsecond_download:
-            log.debug("Checking for updated leapseconds")
-            astropy.time.update_leap_seconds()
-            _leapsecond_download = True
-
     def _load_gps_clock(self):
         global _gps_clock
         if _gps_clock is None:
@@ -303,10 +285,7 @@ class TopoObs(Observatory):
         """
 
         # Read clock file if necessary
-        # and download (if necessary) IERS-B and leapsecond files
         # TODO provide some method for re-reading the clock file?
-        self._load_iersb()
-        self._load_leapsecond()
         self._load_clock_corrections()
         if not self._clock:
             msg = f"No clock corrections found for observatory {self.name} taken from file {self.clock_file}"
